requirements-dev.txt
fastapi
pydantic
uvicorn
langchain
pydantic-settings
anthropic
jupyter
ruff
boto3
google-auth
google-api-python-client
unstructured
opensearch-py
llama-index
html2text
beautifulsoup4
sentence_transformers
lark
requests-aws4auth
elasticsearch